Move SIGWINCH stuff into tty module
